Move User factory to lib/ directory
This way, one can require the user factory from the host app, making
testing complex Alchemy applications easier.
