+++ spec/spec_helper.rb
@@ -21,7 +21,7 @@ require 'alchemy/seeder'
 require 'alchemy/test_support/controller_requests'
 require 'alchemy/test_support/integration_helpers'
 require 'alchemy/test_support/factories'
-require_relative 'support/factories'
+require 'alchemy/devise/test_support/factories'
 
 ActionMailer::Base.delivery_method = :test
 ActionMailer::Base.perform_deliveries = true
